perf: add metadata fetch coverage for tag-aware config diffs
Add unit and functional coverage for skipping tag metadata queries when model
config does not require them. This protects the new fetch-planning logic
across incremental, view, streaming table, and materialized view test paths
without changing unrelated unstaged work.
