+++ tests/functional/adapter/incremental/test_incremental_metadata_fetch_skips.py
@@ -1,17 +1,10 @@
 import pytest
 from dbt.tests import util
 
+from tests.functional.adapter.helpers import (
+    FAIL_IF_TAG_AND_COLUMN_TAG_FETCH_CALLED_MACROS,
+)
 from tests.functional.adapter.incremental import fixtures
-
-FAIL_IF_TAG_FETCH_CALLED_MACROS = """
-{% macro fetch_tags(relation) %}
-  {{ exceptions.raise_compiler_error("fetch_tags should not be called") }}
-{% endmacro %}
-
-{% macro fetch_column_tags(relation) %}
-  {{ exceptions.raise_compiler_error("fetch_column_tags should not be called") }}
-{% endmacro %}
-"""
 
 
 @pytest.mark.skip_profile("databricks_cluster")
@@ -25,9 +18,13 @@ class TestIncrementalMetadataFetchSkips:
 
     @pytest.fixture(scope="class")
     def macros(self):
-        return {"fail_if_tag_fetch_called.sql": FAIL_IF_TAG_FETCH_CALLED_MACROS}
+        return {
+            "fail_if_tag_fetch_called.sql": FAIL_IF_TAG_AND_COLUMN_TAG_FETCH_CALLED_MACROS
+        }
 
     def test_second_incremental_run_succeeds_without_tag_fetches(self, project):
+        # The first run creates the relation; the second run exercises the existing-relation
+        # path where adapter.get_relation_config() may attempt metadata fetches.
         util.run_dbt(["run"])
         util.run_dbt(["run"])
 
@@ -43,9 +40,13 @@ class TestIncrementalMetadataFetchRequiresTableTags:
 
     @pytest.fixture(scope="class")
     def macros(self):
-        return {"fail_if_tag_fetch_called.sql": FAIL_IF_TAG_FETCH_CALLED_MACROS}
+        return {
+            "fail_if_tag_fetch_called.sql": FAIL_IF_TAG_AND_COLUMN_TAG_FETCH_CALLED_MACROS
+        }
 
     def test_second_incremental_run_fails_when_table_tag_fetch_is_required(self, project):
+        # The first run creates the relation; the second run exercises the existing-relation
+        # path where adapter.get_relation_config() may attempt metadata fetches.
         util.run_dbt(["run"])
         _, logs = util.run_dbt_and_capture(["run"], expect_pass=False)
         util.assert_message_in_logs("fetch_tags should not be called", logs)
@@ -62,9 +63,13 @@ class TestIncrementalMetadataFetchRequiresColumnTags:
 
     @pytest.fixture(scope="class")
     def macros(self):
-        return {"fail_if_tag_fetch_called.sql": FAIL_IF_TAG_FETCH_CALLED_MACROS}
+        return {
+            "fail_if_tag_fetch_called.sql": FAIL_IF_TAG_AND_COLUMN_TAG_FETCH_CALLED_MACROS
+        }
 
     def test_second_incremental_run_fails_when_column_tag_fetch_is_required(self, project):
+        # The first run creates the relation; the second run exercises the existing-relation
+        # path where adapter.get_relation_config() may attempt metadata fetches.
         util.run_dbt(["run"])
         _, logs = util.run_dbt_and_capture(["run"], expect_pass=False)
         util.assert_message_in_logs("fetch_column_tags should not be called", logs)
