+++ tests/functional/adapter/materialized_view_tests/test_changes.py
@@ -14,8 +14,8 @@ from dbt.adapters.databricks.relation_configs.materialized_view import (
     MaterializedViewConfig,
 )
 from dbt.adapters.databricks.relation_configs.tblproperties import TblPropertiesConfig
+from tests.functional.adapter.helpers import get_model_config
 from tests.functional.adapter.materialized_view_tests import fixtures
-from tests.functional.adapter.utils.test_utils import get_model_config
 
 
 def _check_tblproperties(tblproperties: TblPropertiesConfig, expected: dict):
